삭제

+++ src/main/java/com/test/project/store/StoreRepository.java
@@ -2,5 +2,5 @@ package com.test.project.store;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreRepository extends JpaRepository<Store, Long> {
+public interface StoreRepository extends JpaRepository<Store, Integer> {
 }
